refactor: move common container functions to utils module
