+++ src/tomodachi_testcontainers/containers/tomodachi.py
@@ -5,7 +5,8 @@ from typing import Any, Optional
 
 from testcontainers.core.waiting_utils import wait_for_logs
 
-from tomodachi_testcontainers.containers.common import DockerContainer, wait_for_http_healthcheck
+from tomodachi_testcontainers.containers.common import DockerContainer
+from tomodachi_testcontainers.utils import wait_for_http_healthcheck
 
 
 class TomodachiContainer(DockerContainer):
