+++ tests/containers/common/test_ephemeral_docker_image.py
@@ -5,7 +5,8 @@ from typing import Generator
 import pytest
 from docker.errors import BuildError, ImageNotFound
 
-from tomodachi_testcontainers import EphemeralDockerImage, get_docker_image
+from tomodachi_testcontainers import EphemeralDockerImage
+from tomodachi_testcontainers.utils import get_docker_image
 
 
 @pytest.fixture()
